integration/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# HTML sanitization
beautifulsoup4==4.12.3
lxml==5.3.0
bleach==6.1.0
html5lib==1.1

# HTTP clients
httpx==0.27.2
aiohttp==3.10.8

# Cache & session
redis[hiredis]==5.1.1

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
psycopg2-binary==2.9.9

# pgvector — vector similarity search
pgvector==0.3.5
numpy==1.26.4

# PII detection
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.6

# ML classifiers
scikit-learn==1.5.2
transformers==4.45.2
torch==2.4.1
sentence-transformers==3.2.0

# Domain/URL analysis
tldextract==5.1.3
dnspython==2.7.0
python-whois==0.9.4

# Unicode / obfuscation
confusables==1.2.0

# Logging
structlog==24.4.0
python-json-logger==2.0.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
factory-boy==3.3.1
freezegun==1.5.1

# Utils
python-multipart==0.0.12
PyYAML==6.0.2
python-dotenv==1.0.1
